requirements.txt
aiogram==3.13.1
motor==3.6.0
python-dotenv==1.0.1